+++ cqrs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # master
 
-* No changes yet
+* Update `cqrs-core` dependency
 
 # [[0.2.1] 2019-03-08](https://github.com/cq-rs/cqrs/releases/tag/cqrs-0.2.1)